refactor: keep transport modules within size policy
