+++ crates/libs/rns-transport/src/transport/tests_parts/held_udp_announce_preserves_peer_sou.rs
@@ -492,14 +492,4 @@ async fn resource_request_responses_fit_bound_iface_mtu() {
     }
 }
 
-fn decrypt_resource_advertisement(link: &Link, packet: &Packet) -> ResourceAdvertisement {
-    let mut buffer = PacketDataBuffer::new();
-    let plain_len = {
-        let plain = link
-            .decrypt(packet.data.as_slice(), buffer.accuire_buf_max())
-            .expect("decrypt resource advertisement");
-        plain.len()
-    };
-    buffer.resize(plain_len);
-    ResourceAdvertisement::unpack(buffer.as_slice()).expect("resource advertisement")
-}
+include!("resource_advertisement_test_helper.rs");
